script: Move a bunch of routines used by codegen to script_bindings.
